+++ tests/testthat/test-lambda-sequence.R
@@ -1,3 +1,16 @@
 test_that("lambda sequences are generated correctly", {
+  for (lambda_type in c("gaussian", "bh", "oscar")) {
+    # n = p
+    lambda1 <- regularizationWeights(10, lambda_type = lambda_type, n = 10)
 
+    # n > p
+    lambda2 <- regularizationWeights(10, lambda_type = lambda_type, n = 100)
+
+    # n < p
+    lambda3 <- regularizationWeights(100, lambda_type = lambda_type, n = 10)
+
+    expect_true(all(diff(lambda1) <= 0))
+    expect_true(all(diff(lambda2) <= 0))
+    expect_true(all(diff(lambda3) <= 0))
+  }
 })